+++ Eudora71/Tests/TestMain.cpp
@@ -25,6 +25,7 @@ void RunE88Tests(void);
 void RunE89Tests(void);
 void RunE101Tests(void);
 void RunResolveURLTests(void);
+void RunE112Tests(void);
 
 //
 // Ermittelt Eudora71\Eudora aus dem Ablageort des Testprogramms.
@@ -108,6 +109,7 @@ int main(int argc, char* argv[])
 	RunE89Tests();
 	RunE101Tests();
 	RunResolveURLTests();
+	RunE112Tests();
 
 	iFailed = TT_Summary();
 	return iFailed;